fix: handle additional login errors
This commit adds additional handling for discord accounts not having
verified emails, and oauth callback errors as well as other sign in
errors. Previously, users were presented with an unhelpful black page
with no link back to the rest of the website. After this commit, users
will instead be redirected back to the log in page to attempt to log in
again. They are also given a helpful error message for common failure
cases.